requirements.txt
flask
request
flask-mysql